+++ polaris.shopify.com/playroom/process-code.ts
@@ -1,0 +1,18 @@
+export default function processCode(code: string) {
+  const codeLines = code.trim().split('\n\n');
+  const jsx = codeLines.pop() || '';
+  return `{
+      (() => {
+        ${codeLines.join('\n\n')}
+        ${
+          jsx.trim()?.startsWith('return')
+            ? jsx
+            : `return (
+            <>
+              ${jsx}
+            </>
+          )`
+        }
+      })()
+  }`;
+}